Use interface instead of implemenations for Shakes
